2.12 version

+++ backend/controllers/NewsController.php
@@ -1,13 +1,22 @@
 <?php
 
+
+/**
+ * Team: "小组"小组
+ * Coding by 2013544
+ * This is the newscontroller of backend controller
+ */
+
 namespace backend\controllers;
 
+use yii\web\ForbiddenHttpException;
 use Yii;
 use common\models\News;
 use common\models\NewsSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\filters\AccessControl;
 
 /**
  * NewsController implements the CRUD actions for News model.
@@ -26,6 +35,22 @@ class NewsController extends Controller
                     'delete' => ['POST'],
                 ],
             ],
+            'access' =>[
+                'class' => AccessControl::className(),
+                'rules' =>
+                [
+                        [
+                                'actions' => ['index', 'view'],
+                                'allow' => true,
+                                'roles' => ['?'],
+                        ],
+                [
+                'actions' => ['view', 'index', 'create','update','delete'],
+                'allow' => true,
+                'roles' => ['@'],
+                ],
+                ],
+            ],        
         ];
     }
 
@@ -64,6 +89,9 @@ class NewsController extends Controller
      */
     public function actionCreate()
     {
+        if(!Yii::$app->user->can('createPost')) {
+            throw new ForbiddenHttpException('对不起，您没有进行该操作的权限');
+        }
         $model = new News();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -84,6 +112,9 @@ class NewsController extends Controller
      */
     public function actionUpdate($id)
     {
+        if(!Yii::$app->user->can('updatePost')) {
+            throw new ForbiddenHttpException('对不起，您没有进行该操作的权限');
+        }
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -104,6 +135,9 @@ class NewsController extends Controller
      */
     public function actionDelete($id)
     {
+        if(!Yii::$app->user->can('deletePost')) {
+            throw new ForbiddenHttpException('对不起，您没有进行该操作的权限');
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
